新增积分信息管理接口与其功能方法实现

+++ rcc-user8002/src/main/java/com/aenlly/rcc/user/service/IPointsService.java
@@ -24,5 +24,5 @@ public interface IPointsService extends IService<Points> {
    * @param points 累积积分
    * @return 当前等级
    */
-  Points getCurrentLevel(Integer points);
+  Points getCurrentLevel(Long points);
 }
